feat(NumberInput): Replace deprecated implementation
