+++ packages/react-component-library/src/components/NumberInput/Input.tsx
@@ -1,50 +1,37 @@
 import React from 'react'
-import { isFinite, isNil } from 'lodash'
-import { selectors } from '@defencedigital/design-tokens'
+import { isNil } from 'lodash'
 
-import { NumberInputUnit } from './NumberInputUnit'
-import { UnitPosition } from './NumberInput'
-import { useInputText } from './useInputText'
-import { InputValidationProps } from '../../common/InputValidationProps'
-import { StyledInput } from './partials/StyledInput'
-import { StyledLabel } from './partials/StyledLabel'
+import { ComponentSizeType } from '../Forms'
+import { StyledInput } from '../TextInputE/partials/StyledInput'
 import { StyledInputWrapper } from './partials/StyledInputWrapper'
+import { StyledLabel } from '../TextInputE/partials/StyledLabel'
 
-export interface InputProps extends InputValidationProps {
+export interface InputProps {
   hasFocus: boolean
   isDisabled?: boolean
   id?: string
-  isCondensed: boolean
   label?: string
   name: string
+  onBeforeInput: (event: React.FormEvent<HTMLInputElement>) => void
   onBlur: (event: React.FormEvent<HTMLInputElement>) => void
   onChange: (event: React.ChangeEvent<HTMLInputElement>) => void
   onFocus: (event: React.FormEvent<HTMLInputElement>) => void
+  onPaste: (event: React.ClipboardEvent<HTMLInputElement>) => void
   placeholder?: string
-  unit: string
-  unitPosition: UnitPosition
-  value?: number
+  size: ComponentSizeType
+  value?: string | null
 }
-
-const { spacing } = selectors
 
 export const Input: React.FC<InputProps> = ({
   hasFocus,
-  isDisabled = false,
+  isDisabled,
   id,
-  isCondensed,
   label,
-  placeholder = '',
-  unit,
-  unitPosition,
+  size,
   value,
   ...rest
 }) => {
   const hasLabel = !!(label && label.length)
-  const { canShow, inputOffset, inputRef, unitOffset } = useInputText(
-    value,
-    unitPosition
-  )
 
   return (
     <StyledInputWrapper>
@@ -52,34 +39,20 @@ export const Input: React.FC<InputProps> = ({
         <StyledLabel
           $hasContent={!isNil(value)}
           $hasFocus={hasFocus}
-          $hasUnit={!isNil(unit)}
-          data-testid="number-input-label"
+          $size={size}
           htmlFor={id}
+          data-testid="number-input-label"
         >
           {label}
         </StyledLabel>
       )}
-
-      {canShow && (
-        <NumberInputUnit
-          left={`${unitOffset}px`}
-          top={hasLabel && spacing('4')}
-        >
-          {unit}
-        </NumberInputUnit>
-      )}
-
       <StyledInput
         $hasLabel={hasLabel}
-        $isCondensed={isCondensed}
-        $offset={inputOffset}
+        $size={size}
         data-testid="number-input-input"
         disabled={isDisabled}
         id={id}
-        placeholder={placeholder}
-        ref={inputRef}
-        type="text"
-        value={isFinite(value) && canShow ? value : ''}
+        value={value || ''}
         {...rest}
       />
     </StyledInputWrapper>
