+++ packages/react-component-library/src/components/NumberInput/NumberInput.tsx
@@ -1,36 +1,32 @@
 import React from 'react'
-import { isFinite, isNil } from 'lodash'
+import { isNil } from 'lodash'
 import { v4 as uuidv4 } from 'uuid'
 
-import { ClearButton } from './ClearButton'
-import { EndAdornment } from './EndAdornment'
-import { Footnote } from './Footnote'
+import { Buttons } from './Buttons'
+import { COMPONENT_SIZE, ComponentSizeType } from '../Forms'
 import { getId, hasClass } from '../../helpers'
 import { Input } from './Input'
 import { InputValidationProps } from '../../common/InputValidationProps'
 import { ComponentWithClass } from '../../common/ComponentWithClass'
-import { StartAdornment } from './StartAdornment'
+import { useChangeHandlers } from './useChangeHandlers'
 import { useFocus } from '../../hooks/useFocus'
+import { useEarlyValidation } from './useEarlyValidation'
 import { useValue } from './useValue'
-import { UNIT_POSITION } from './constants'
+import { StyledIcon } from './partials/StyledIcon'
+import { StyledDivider } from './partials/StyledDivider'
+import { StyledFootnote } from './partials/StyledFootnote'
 import { StyledNumberInput } from './partials/StyledNumberInput'
-import { StyledNumberInputOuterWrapper } from './partials/StyledNumberInputOuterWrapper'
+import { StyledOuterWrapper } from './partials/StyledOuterWrapper'
+import { StyledPrefix } from './partials/StyledPrefix'
+import { StyledSuffix } from './partials/StyledSuffix'
 
-export type UnitPosition =
-  | typeof UNIT_POSITION.AFTER
-  | typeof UNIT_POSITION.BEFORE
-
-export interface NumberInputProps
+interface NumberInputBaseProps
   extends ComponentWithClass,
     InputValidationProps {
   /**
    * Toggles whether to focus the input on initial render.
    */
   autoFocus?: boolean
-  /**
-   * Toggles whether or not the clear button is displayed.
-   */
-  canClear?: boolean
   /**
    * Optional text footnote to display below the component.
    */
@@ -39,10 +35,6 @@ export interface NumberInputProps
    * Optional HTML `id` attribute to apply to the internal input.
    */
   id?: string
-  /**
-   * Toggles whether or not to render a reduced height version of the component.
-   */
-  isCondensed?: boolean
   /**
    * Toggles whether the component is disabled or not (preventing user interaction).
    */
@@ -57,6 +49,9 @@ export interface NumberInputProps
   max?: number
   /**
    * Minimum value selectable by the component (lower bound).
+   *
+   * To disable the entry of negative numbers, set this to `0` or
+   * greater.
    */
   min?: number
   /**
@@ -68,76 +63,92 @@ export interface NumberInputProps
    */
   onBlur?: (event: React.FormEvent) => void
   /**
-   * Handler called when the value selected by the component changes.
+   * Handler called when the entered value changes.
+   *
+   * Note that `newValue` could be `NaN` if negative numbers
+   * are allowed and a single `-` is entered.
    */
-  onChange: (event: any) => void
+  onChange: (
+    event:
+      | React.ChangeEvent<HTMLInputElement>
+      | React.MouseEvent<HTMLButtonElement>,
+    newValue: number | null
+  ) => void
   /**
    * Optional placeholder text to display within the component.
    */
   placeholder?: string
   /**
-   * Optional adornment (icon or string) to display to the left of the input value.
+   * Size of the component.
    */
-  startAdornment?: React.ReactNode | string
+  size?: ComponentSizeType
   /**
    * Stepped increment amount by which to increase/decrese component value.
    */
   step?: number
   /**
-   * Optional value to display next to the component value.
-   */
-  unit?: string
-  /**
-   * Whether to display the unit to the left or right of the component value.
-   */
-  unitPosition?: UnitPosition
-  /**
    * Currently selected value displayed by the component.
    */
-  value?: number
+  value?: number | null
 }
 
+export interface NumberInputWithIconProps extends NumberInputBaseProps {
+  /**
+   * Optional icon to display to the left of the input value.
+   */
+  icon?: React.ReactNode
+  prefix?: never
+  suffix?: never
+}
+
+export interface NumberInputWithPrefixProps extends NumberInputBaseProps {
+  icon?: never
+  /**
+   * Optional value to display next to the component value.
+   */
+  prefix?: string
+  suffix?: never
+}
+
+export interface NumberInputWithSuffixProps extends NumberInputBaseProps {
+  icon?: never
+  prefix?: never
+  /**
+   * Optional value to display next to the component value.
+   */
+  suffix?: string
+}
+
+export type NumberInputProps =
+  | NumberInputWithIconProps
+  | NumberInputWithPrefixProps
+  | NumberInputWithSuffixProps
+
 function formatValue(
-  displayValue: number,
-  unit: string,
-  unitPosition: UnitPosition
-) {
+  displayValue: string | null,
+  prefix?: string,
+  suffix?: string
+): string {
   if (isNil(displayValue)) {
-    return null
+    return 'Not set'
   }
 
-  if (unit) {
-    return unitPosition === UNIT_POSITION.AFTER
-      ? `${displayValue} ${unit}`
-      : `${unit} ${displayValue}`
+  if (prefix) {
+    return `${prefix} ${displayValue}`
+  }
+
+  if (suffix) {
+    return `${displayValue} ${suffix}`
   }
 
   return displayValue
 }
 
-function getNewValue(event: React.FormEvent<HTMLInputElement>): number {
-  const { value } = event.currentTarget
-
-  if (value === '') {
-    return null
-  }
-
-  return parseInt(value, 10)
-}
-
-function isWithinRange(max: number, min: number, newValue: number) {
-  const isNotBelowMin = isNil(min) || newValue >= min
-  const isNotAboveMax = isNil(max) || newValue <= max
-
-  return isNotBelowMin && isNotAboveMax
-}
-
 export const NumberInput: React.FC<NumberInputProps> = ({
-  canClear,
   className,
   footnote,
+  icon,
   id = uuidv4(),
-  isCondensed,
   isDisabled = false,
   isInvalid,
   label,
@@ -147,103 +158,112 @@ export const NumberInput: React.FC<NumberInputProps> = ({
   onBlur,
   onChange,
   placeholder = '',
-  startAdornment,
+  prefix,
+  size = COMPONENT_SIZE.FORMS,
   step = 1,
-  unit,
-  unitPosition = UNIT_POSITION.AFTER,
-  value = 0,
+  suffix,
+  value = null,
   ...rest
 }) => {
-  const { committedValue, setCommittedValue } = useValue(value)
+  const isNegativeAllowed = isNil(min) || min < 0
+  const { committedValue, setCommittedValue } = useValue(
+    value ? String(value) : null
+  )
   const { hasFocus, onLocalFocus, onLocalBlur } = useFocus(onBlur)
-
-  function setCommittedValueWithinRange(newValue: number) {
-    if (
-      (isFinite(newValue) && isWithinRange(max, min, newValue)) ||
-      newValue === null
-    ) {
-      setCommittedValue(newValue)
-      onChange({
-        target: {
-          name,
-          value: newValue,
-        },
-      })
-    }
-  }
+  const { handleBeforeInput, handlePaste } =
+    useEarlyValidation(isNegativeAllowed)
+  const { handleButtonClick, handleInputChange } = useChangeHandlers(
+    isNegativeAllowed,
+    min,
+    max,
+    onChange,
+    setCommittedValue
+  )
 
   const numberInputId = getId('number-input')
+  const isCommittedValueInvalid =
+    committedValue && !Number.isFinite(parseFloat(committedValue))
 
   return (
     <StyledNumberInput
       aria-label={label || 'Number input'}
       className={className}
-      data-testid="number-input-container"
+      data-testid="number-input"
       id={numberInputId}
       role="spinbutton"
       aria-valuemin={min}
       aria-valuemax={max}
-      aria-valuenow={committedValue}
-      aria-valuetext={String(formatValue(committedValue, unit, unitPosition))}
+      aria-valuenow={Number(committedValue) || 0}
+      aria-valuetext={String(formatValue(committedValue, prefix, suffix))}
     >
-      <StyledNumberInputOuterWrapper
+      <StyledOuterWrapper
+        data-testid="number-input-outer-wrapper"
         $hasFocus={hasFocus}
-        $isInvalid={isInvalid || hasClass(className, 'is-invalid')}
+        $isDisabled={isDisabled}
+        $isInvalid={
+          isInvalid ||
+          isCommittedValueInvalid ||
+          hasClass(className, 'is-invalid')
+        }
+        $size={size}
       >
-        <StartAdornment>{startAdornment}</StartAdornment>
+        {icon && (
+          <StyledIcon data-testid="number-input-icon">{icon}</StyledIcon>
+        )}
+
+        {prefix && (
+          <>
+            <StyledPrefix data-testid="number-input-prefix">
+              {prefix}
+            </StyledPrefix>
+            <StyledDivider $size={size} />
+          </>
+        )}
 
         <Input
           aria-labelledby={numberInputId}
           hasFocus={hasFocus}
           id={id}
           isDisabled={isDisabled}
-          isCondensed={isCondensed}
-          isInvalid={isInvalid || hasClass(className, 'is-invalid')}
           label={label}
           name={name}
-          onChange={(event) => {
-            const newValue = getNewValue(event)
-            setCommittedValueWithinRange(newValue)
-          }}
-          onBlur={(event) => {
-            const newValue = getNewValue(event)
-            setCommittedValueWithinRange(newValue)
-
-            onLocalBlur(event)
-          }}
+          onBeforeInput={handleBeforeInput}
+          onChange={handleInputChange}
+          onPaste={handlePaste}
+          onBlur={onLocalBlur}
           onFocus={onLocalFocus}
           placeholder={placeholder}
-          unit={unit}
-          unitPosition={unitPosition}
+          size={size}
           value={committedValue}
           {...rest}
         />
 
-        {canClear && (
-          <ClearButton
-            isCondensed={isCondensed}
-            isDisabled={isDisabled}
-            onClick={() => {
-              setCommittedValueWithinRange(null)
-            }}
-          />
+        {suffix && (
+          <>
+            <StyledDivider $size={size} />
+            <StyledSuffix data-testid="number-input-suffix">
+              {suffix}
+            </StyledSuffix>
+          </>
         )}
 
-        <EndAdornment
-          isCondensed={isCondensed}
+        <Buttons
           isDisabled={isDisabled}
           max={max}
           min={min}
           name={name}
-          onClick={(e, newValue) => {
-            setCommittedValueWithinRange(newValue)
-          }}
+          onClick={handleButtonClick}
+          size={size}
           step={step}
           value={committedValue}
         />
-      </StyledNumberInputOuterWrapper>
+      </StyledOuterWrapper>
 
-      <Footnote>{footnote}</Footnote>
+      {footnote && (
+        <StyledFootnote data-testid="number-input-footnote">
+          {footnote}
+        </StyledFootnote>
+      )}
     </StyledNumberInput>
   )
 }
