+++ packages/react-component-library/src/components/NumberInput/partials/StyledDivider.tsx
