+++ packages/react-component-library/src/components/NumberInput/partials/StyledFootnote.tsx
