+++ packages/react-component-library/src/components/NumberInput/partials/StyledIcon.tsx
