--- packages/react-component-library/src/components/NumberInput/partials/StyledNumberInputControls.tsx
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-import { selectors } from '@defencedigital/design-tokens'
-
-const { color } = selectors
-
-export const StyledNumberInputControls = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  text-align: center;
-  border-left: 1px solid ${color('neutral', '100')};
-`
