+++ packages/react-component-library/src/components/NumberInput/partials/StyledOuterWrapper.tsx
