+++ packages/react-component-library/src/components/NumberInput/partials/StyledPrefix.tsx
