+++ packages/react-component-library/src/components/NumberInput/partials/StyledSuffix.tsx
