+++ packages/react-component-library/src/components/NumberInput/useEarlyValidation.ts
