+++ packages/react-component-library/src/components/NumberInput/useValue.ts
@@ -1,6 +1,9 @@
-import { useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 
-export function useValue(value: number) {
+export function useValue(value: string | null): {
+  committedValue: string | null
+  setCommittedValue: React.Dispatch<React.SetStateAction<string | null>>
+} {
   const [committedValue, setCommittedValue] = useState(value)
 
   useEffect(() => {
