+++ packages/react-component-library/src/components/NumberInput/validation.ts
