--- packages/react-component-library/src/components/NumberInputE/constants.ts
@@ -1,6 +1,0 @@
-const NUMBER_INPUT_BUTTON_TYPE = {
-  DECREASE: 'decrease',
-  INCREASE: 'increase',
-} as const
-
-export { NUMBER_INPUT_BUTTON_TYPE }
