--- packages/react-component-library/src/components/NumberInputE/index.ts
@@ -1,2 +1,0 @@
-export * from './constants'
-export * from './NumberInputE'
