--- packages/react-component-library/src/components/NumberInputE/partials/StyledInputWrapper.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledInputWrapper = styled.div`
-  position: relative;
-  flex-grow: 1;
-`
