--- packages/react-component-library/src/components/NumberInputE/partials/StyledNumberInput.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledNumberInput = styled.div`
-  display: inline-flex;
-  flex-direction: column;
-  position: relative;
-  padding: 0;
-  border: 0;
-  vertical-align: top;
-  width: 100%;
-`
