--- packages/react-component-library/src/components/NumberInputE/useValue.ts
@@ -1,17 +1,0 @@
-import React, { useEffect, useState } from 'react'
-
-export function useValue(value: string | null): {
-  committedValue: string | null
-  setCommittedValue: React.Dispatch<React.SetStateAction<string | null>>
-} {
-  const [committedValue, setCommittedValue] = useState(value)
-
-  useEffect(() => {
-    setCommittedValue(value)
-  }, [value])
-
-  return {
-    committedValue,
-    setCommittedValue,
-  }
-}
