+++ packages/react-component-library/src/forms/native/native.stories.tsx
@@ -8,7 +8,7 @@ import { Radio } from '../../components/Radio'
 import { Autocomplete, AutocompleteOption } from '../../components/Autocomplete'
 import { Checkbox } from '../../components/Checkbox'
 import { DatePicker } from '../../components/DatePicker'
-import { NumberInputE } from '../../components/NumberInputE'
+import { NumberInput } from '../../components/NumberInput'
 import { SelectE, SelectEOption } from '../../components/SelectE'
 import { SwitchE, SwitchEOption } from '../../components/SwitchE'
 import { RangeSliderE } from '../../components/RangeSliderE'
@@ -170,7 +170,7 @@ export const Example: React.FC<unknown> = () => {
           </SwitchE>
         </Field>
         <Field hintText="Example hint text.">
-          <NumberInputE
+          <NumberInput
             name="exampleNumberInput"
             label="Example number input"
             onChange={(
@@ -187,7 +187,7 @@ export const Example: React.FC<unknown> = () => {
               })
             }}
             value={formState.exampleNumberInput}
-            data-testid="form-example-NumberInputE"
+            data-testid="form-example-NumberInput"
           />
         </Field>
         <Field
