+++ packages/react-component-library/src/index.ts
@@ -32,7 +32,6 @@ export * from './components/Link'
 export * from './components/List'
 export * from './components/Modal'
 export * from './components/NumberInput'
-export * from './components/NumberInputE'
 export * from './components/Pagination'
 export * from './components/Panel'
 export * from './components/PhaseBanner'
